Harvester metrics wrong when the host is repeated
When multiple harvesters have the same domain name or IP (when behind
a router) we only get the last harvester in the metric, because we keep
replacing the earlier ones.
By appending the node_id to the host they become unique and correctly
reported. As a bonus those node IDs are the same ones identifying
the harvesters in flexpool so you can track which is which.